\name{EGQuery}
\alias{EGQuery}
\title{New EGQuery Object}
\usage{EGQuery(NCBIObject)}
\description{This function creates an object of the EGQuery class and returns it.}
\value{An object of the EGQuery class.
#'}
\seealso{\code{\link[=EGQueryClass-class]{EGQueryClass}}}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}}
