\name{EGQueryClass}
\title{Represents The EGQuery Operation}
\description{This class represents the EGQuery operation. It contains NCBIResult and NCBIRequest to hold the results and the request respectively.}
\seealso{\code{\link{requestEGQuery}}}
\alias{EGQueryClass-class}
\author{Martin Schumann}
