\name{EInfo}
\alias{EInfo}
\title{New EInfo Object}
\usage{EInfo(NCBIObject)}
\description{This function creates an object of the EInfo class and returns it.}
\value{An object of the EInfo class.}
\seealso{\code{\link[=EInfoClass-class]{EInfoClass}}}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass. The instance of the NCBI class. Required to get access to the java interface.}}
