\name{EInfoClass}
\title{Represents The EInfo Operation}
\description{This class represents the EInfo operation. It contains NCBIResult and NCBIRequest to hold the results and the request respectively.
The EInfo operation contains a logical value to indicate whether the returned result is a DBInfo or not.}
\seealso{\code{\link{requestEInfo}}}
\alias{EInfoClass-class}
\author{Martin Schumann}
