\name{ELink}
\alias{ELink}
\title{New ELink Object}
\usage{ELink(NCBIObject)}
\description{This function creates an object of the ELink class and returns it.}
\value{An object of the ELink class.}
\seealso{\code{\link[=ELinkClass-class]{ELinkClass}}}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}}
