\name{EPost}
\alias{EPost}
\title{New EPost Object}
\usage{EPost(NCBIObject)}
\description{This function creates an object of the EPost class and returns it.}
\value{An object of the EPost class.}
\seealso{\code{\link[=EPostClass-class]{EPostClass}}}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}}
