\name{ESearch}
\alias{ESearch}
\title{New ESearch Object}
\usage{ESearch(NCBIObject)}
\description{This function creates an object of the ESearch class and returns it.}
\value{An object of the ESearch class.
#'}
\seealso{\code{\link[=ESearchClass-class]{ESearchClass}}}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}}
