\name{ESpell}
\alias{ESpell}
\title{New ESpell Object}
\usage{ESpell(NCBIObject)}
\description{This function creates an object of the ESpell class and returns it.}
\value{An object of the ESpell class.}
\seealso{\code{\link[=ESpellClass-class]{ESpellClass}}}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}}
