\name{ESummary}
\alias{ESummary}
\title{New ESummary Object}
\usage{ESummary(NCBIObject)}
\description{This function creates an object of the ESummary class and returns it.}
\value{An object of the ESummary class.}
\seealso{\code{\link[=ESummaryClass-class]{ESummaryClass}}}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}}
