\name{.getComplexType}
\title{Complex Type}
\usage{.getComplexType(opObj, parName)}
\description{This will retrieve the complex type identified by "parName" from
the java interface.}
\value{The same object as the arguments, but with a modified results list.}
\alias{.getComplexType,NCBIResult-method}
\alias{.getComplexType}
\author{Martin Schumann}
\arguments{\item{opObj}{An object of this class.}
\item{parName}{The parameter name of the complex type, which should be retrieved.}}
