\name{getEFetchDatabases}
\title{Get All EFetch Database Names}
\usage{getEFetchDatabases(NCBIObject, printNames=TRUE)}
\description{Retrieves all available databases for the EFetch operation from the java interface.}
\value{A vector with the names of the operation or nothing.}
\alias{getEFetchDatabases,NCBIClass-method}
\alias{getEFetchDatabases}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}
\item{printNames}{Whether to print or to return a vector.}}
