\name{.getEInfoResults}
\title{Parse The Results}
\usage{.getEInfoResults(EInfoObj)}
\description{Parses the results parameter names and calls the appropriate function of \code{\link[=NCBIResult-class]{NCBIResult}}, 
depending on whether the parameter is a complex type or a simple type.}
\value{The object with the results, if no errors occured.}
\alias{.getEInfoResults,EInfoClass-method}
\alias{.getEInfoResults}
\author{Martin Schumann}
\arguments{\item{EInfoObj}{An object of this class.}}
