\name{.getELinkResults}
\title{Parse The Results}
\usage{.getELinkResults(ELObj)}
\description{Parses the results parameter names and calls the appropriate function of \code{\link[=NCBIResult-class]{NCBIResult}}, 
depending on whether the parameter is a complex type or a simple type.}
\value{The modified object with the results, if no errors occured.}
\alias{.getELinkResults,ELinkClass-method}
\alias{.getELinkResults}
\author{Martin Schumann}
\arguments{\item{ELObj}{An object of this class.}}
