\name{.getEPostResults}
\title{Parse The Results}
\usage{.getEPostResults(EPObj)}
\description{Parses the results parameter names and calls the appropriate function of \code{\link[=NCBIResult-class]{NCBIResult}}, 
depending on whether the parameter is a complex type or a simple type.}
\value{The object containing the results, if no errors occured.}
\alias{.getEPostResults,EPostClass-method}
\alias{.getEPostResults}
\author{Martin Schumann}
\arguments{\item{EPObj}{An object of this class.}}
