\name{.getESearchResults}
\title{Parse The Results}
\usage{.getESearchResults(ESObj)}
\description{Parses the results parameter names and calls the appropriate function of \code{\link[=NCBIResult-class]{NCBIResult}}, 
depending on whether the parameter is a complex type or a simple type.}
\value{The object with the results, if no errors occured.}
\alias{.getESearchResults,ESearchClass-method}
\alias{.getESearchResults}
\author{Martin Schumann}
\arguments{\item{ESObj}{An object of this class.}}
