\name{.getESummaryResults}
\title{Parse The Results}
\usage{.getESummaryResults(ESObj)}
\description{Parses the results parameter names and calls the appropriate function of \code{\link[=NCBIResult-class]{NCBIResult}}, 
depending on whether it is a complex type or a simple type.}
\value{The object with the results, if no errors occured.}
\alias{.getESummaryResults,ESummaryClass-method}
\alias{.getESummaryResults}
\author{Martin Schumann}
\arguments{\item{ESObj}{An object of this class.}}
