\name{getOperations}
\title{Available EUtils Operations}
\usage{getOperations(NCBIObject, printNames=TRUE)}
\description{Get a list of available operations from the java interface.}
\value{Either a vector with the names of the operations or nothing.}
\alias{getOperations,NCBIClass-method}
\alias{getOperations}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}
\item{printNames}{Whether to print the names or return a vector.}}
