\name{getRequestParameter}
\title{Get Parameter For The Request}
\usage{getRequestParameter(NCBIObject, req="")}
\description{Retrieve a list of request parameter names for the given operation.}
\value{A vector with the request parameter names for the given function.}
\alias{getRequestParameter,NCBIClass-method}
\alias{getRequestParameter}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}
\item{req}{The name of the operation the parameter names should be retrieved.}}
