\name{getResults}
\title{The Results}
\usage{getResults(obj)}
\description{Returns the results of the given operation object.}
\value{The results list from the given operation object.}
\alias{getResults,NCBIResult-method}
\alias{getResults}
\author{Martin Schumann}
\arguments{\item{obj}{An object of the operation classes.}}
