\name{.getSimpleType}
\title{Simple Type}
\usage{.getSimpleType(opObj, parName)}
\description{This will retrieve the simple type identified by "parName" from
the java interface.}
\value{The object from the arguments, but with modified results list.}
\alias{.getSimpleType,NCBIResult-method}
\alias{.getSimpleType}
\author{Martin Schumann}
\arguments{\item{opObj}{Object if this class.}
\item{parName}{The parameter name for the simple type to retrieve.}}
