\name{NCBIRequest}
\title{Represents The Request To The NCBI Webservice
EFetch operation, this list has to be initiated after the database was specified.}
\description{This class represents the request to the NCBI webservice. This is a simple list, which names are the request parameter for each operation. In case of the}
\alias{NCBIRequest-class}
\author{Martin Schumann}
