\name{requestEFetch}
\alias{requestEFetch}
\title{Send Request}
\usage{requestEFetch(efetch)}
\description{Send an EFetch request to the NCBI Webservice. The EFetch object should have been prepared before making any requests.}
\value{The EFetch object with the results inside.}
\author{Martin Schumann}
\arguments{\item{efetch}{The initiated EFetch object.}}
