\name{.requestEFetchMethod}
\title{Calls the appropriate interface method for this operation and stores the results.}
\usage{.requestEFetchMethod(EFObj)}
\description{Calls the appropriate interface method for this operation and stores the results.}
\section{Warning}{
This function should not be called by the user.
}
\value{The modified object from the arguments, if no errors occured.}
\seealso{\code{\link[=requestEFetch]{requestEFetch}}}
\alias{.requestEFetchMethod,EFetchClass-method}
\alias{.requestEFetchMethod}
\author{Martin Schumann}
\arguments{\item{EFObj}{An object of this class}}
