\name{requestEGQuery}
\alias{requestEGQuery}
\title{Send Request}
\usage{requestEGQuery(egquery)}
\description{Send an EGQuery request to the NCBI Webservice. The EGQuery object should have been prepared before making any requests.}
\value{The EGQuery object with the results inside.}
\author{Martin Schumann}
\arguments{\item{egquery}{The initiated EGQuery object.}}
