\name{.requestEGQueryMethod}
\title{Calls the appropriate interface method for this operation and stores the results.}
\usage{.requestEGQueryMethod(EGQObj)}
\description{Calls the appropriate interface method for this operation and stores the results.}
\section{Warning}{
This function should not be called by the user.
}
\value{The provided object with the results, if no errors occured.}
\seealso{\code{\link[=requestEGQuery]{requestEGQuery}}}
\alias{.requestEGQueryMethod,EGQueryClass-method}
\alias{.requestEGQueryMethod}
\author{Martin Schumann}
\arguments{\item{EGQObj}{An object of this class.}}
