\name{requestEInfo}
\alias{requestEInfo}
\title{Send Request}
\usage{requestEInfo(einfo)}
\description{Send an EInfo request to the NCBI Webservice. The EInfo object should have been prepared before making any requests.}
\value{The EInfo object with the results inside.}
\author{Martin Schumann}
\arguments{\item{einfo}{The initiated EInfo object.}}
