\name{.requestEInfoMethod}
\title{Calls the appropriate interface method for this operation and stores the results.}
\usage{.requestEInfoMethod(EInfoObj)}
\description{Calls the appropriate interface method for this operation and stores the results.}
\section{Warning}{
This function should not be called by the user.
}
\value{The object with the results, if no errors occured.}
\seealso{\code{\link[=requestEInfo]{requestEInfo}}}
\alias{.requestEInfoMethod,EInfoClass-method}
\alias{.requestEInfoMethod}
\author{Martin Schumann}
\arguments{\item{EInfoObj}{An object of this class.}}
