\name{requestELink}
\alias{requestELink}
\title{Send Request}
\usage{requestELink(elink)}
\description{Send an ELink request to the NCBI Webservice. The ELink object should have been prepared before making any requests.}
\value{The ELink object with the results inside.}
\author{Martin Schumann}
\arguments{\item{elink}{The initiated ELink object.}}
