\name{.requestELinkMethod}
\title{Calls the appropriate interface method for this operation and stores the results.}
\usage{.requestELinkMethod(ELObj)}
\description{Calls the appropriate interface method for this operation and stores the results.}
\section{Warning}{
This function should not be called by the user.
}
\value{The modified object with the results, if no errors occured.}
\seealso{\code{\link[=requestELink]{requestELink}}}
\alias{.requestELinkMethod,ELinkClass-method}
\alias{.requestELinkMethod}
\author{Martin Schumann}
\arguments{\item{ELObj}{An object of this class.}}
