\name{requestEPost}
\alias{requestEPost}
\title{Send Request}
\usage{requestEPost(epost)}
\description{Send an EPost request to the NCBI Webservice. The EPost object should have been prepared before making any requests.}
\value{The EPost object with the results inside.}
\author{Martin Schumann}
\arguments{\item{epost}{The initiated EPost object.}}
