\name{.requestEPostMethod}
\title{Calls the appropriate interface method for this operation and stores the results.}
\usage{.requestEPostMethod(EPObj)}
\description{Calls the appropriate interface method for this operation and stores the results.}
\section{Warning}{
This function should not be called by the user.
}
\value{The object with the results, if no errors occured.}
\seealso{\code{\link[=requestEPost]{requestEPost}}}
\alias{.requestEPostMethod,EPostClass-method}
\alias{.requestEPostMethod}
\author{Martin Schumann}
\arguments{\item{EPObj}{An object of this class.}}
