\name{requestESearch}
\alias{requestESearch}
\title{Send Request}
\usage{requestESearch(esearch)}
\description{Send an ESearch request to the NCBI Webservice. The ESearch object should have been prepared before making any requests.}
\value{The ESearch object with the results inside.}
\author{Martin Schumann}
\arguments{\item{esearch}{The initiated ESearch object.}}
