\name{requestESpell}
\alias{requestESpell}
\title{Send Request}
\usage{requestESpell(espell)}
\description{Send an ESpell request to the NCBI Webservice. The ESpell object should have been prepared before making any requests.}
\value{The ESpell object with the results inside.}
\author{Martin Schumann}
\arguments{\item{espell}{The initiated ESpell object.}}
