\name{.requestESpellMethod}
\title{Calls the appropriate interface method for this operation and stores the results.}
\usage{.requestESpellMethod(ESObj)}
\description{Calls the appropriate interface method for this operation and stores the results.}
\section{Warning}{
This function should not be called by the user.
}
\value{The object with the results, if no errors occured.}
\seealso{\code{\link[=requestESpell]{requestESpell}}}
\alias{.requestESpellMethod,ESpellClass-method}
\alias{.requestESpellMethod}
\author{Martin Schumann}
\arguments{\item{ESObj}{An object of this class.}}
