\name{requestESummary}
\alias{requestESummary}
\title{Send Request}
\usage{requestESummary(esummary)}
\description{Send an ESummary request to the NCBI Webservice. The ESummary object should have been prepared before making any requests.}
\value{The ESummary object with the results inside.}
\author{Martin Schumann}
\arguments{\item{esummary}{The initiated ESummary object.}}
