\name{RNCBI}
\title{The Java NCBI Interface To R}
\description{The java ncbi interface to R.}
\details{\tabular{ll}{
Package: \tab RNCBI\cr
Type: \tab Package\cr
Version: \tab 0.9.1\cr
Date: \tab 2010-09-19\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
This is an alpha release.

Uses the Axis2 framework from apache to send requests to the NCBI webservice and get results respectively. 
The R part will convert the requests and results to pass them to the java part.
Interpretation of the results is the task of the user.}
\alias{RNCBI}
\docType{package}
\author{Martin Schumann \email{martin.schumann.email@gmail.com}}
\references{https://code.google.com/p/rncbi/}
\keyword{package}
