\name{setProxyUser}
\title{Setting The User And Password}
\usage{setProxyUser(NCBIObject, user, password="")}
\description{This will set the user name and password for the proxy.}
\alias{setProxyUser,NCBIClass-method}
\alias{setProxyUser}
\author{Martin Schumann}
\arguments{\item{NCBIObject}{The current instance of the NCBIClass.}
\item{user}{The username.}
\item{password}{The password.}}
