\name{setRequestParameter}
\title{This method will set the given request parameter and value to the request list.}
\usage{setRequestParameter(opObj, parameter, value)}
\description{This method will set the given request parameter and value to the request list.}
\value{The object from the arguments, but with modified request list.}
\alias{setRequestParameter,NCBIRequest-method}
\alias{setRequestParameter}
\author{Martin Schumann}
\arguments{\item{opObj}{An object of this class.}
\item{parameter}{The parameter name, which should be set.}
\item{value}{The value to set.}}
